Fix create bug
