--- Server/Infrastructure/Repositories/Contracts/IFeedRepository.cs
@@ -1,13 +1,0 @@
-﻿using RssFeeder.Server.Infrastructure.Dto;
-
-namespace RssFeeder.Server.Infrastructure.Repositories.Contracts
-{
-    public interface IFeedRepository
-    {
-        public Task<List<FeedDto>> GetAllFeeds(CancellationToken cancellationToken);
-        public Task UpdateFeed(FeedDto feed, CancellationToken cancellationToken);
-        public Task<bool> ResetFeedDefault(List<Guid> ids, CancellationToken cancellationToken);
-        public Task<bool> DeleteFeed(Guid feedId, CancellationToken cancellationToken);
-    }
-}
-
