+++ Server/Infrastructure/Services/Contracts/IFeedNavigationGroupService.cs
@@ -4,7 +4,7 @@ using RssFeeder.Shared.Model;
 
 namespace RssFeeder.Server.Infrastructure.Services.Contracts;
 
-public interface IFeedGroupService
+public interface IFeedNavigationGroupService
 {
     public Task<List<FeedNavigationGroup>> GetGroupedFeeds(CancellationToken cancellationToken);
     public Task<OneOf<Guid, List<ValidationFailure>>> InsertGroup(FeedNavigationGroup newFeedGroup, CancellationToken cancellationToken);
